Very important please read description
Fixed a ton of PHP and SQL bugs. Please use this setup from now on.

Got rid of Billing; it's kind of useless since it only has like two
attributes and isn't used by anything else.

There are numbers on the SQL files now, that's the order you should
execute the scripts in so Oracle doesn't get mad at you.